refactor: re-sorting trpc related files
