+++ src/components/provider/trpc-provider.tsx
@@ -1,21 +1,22 @@
 "use client"
 
-import SuperJSON from 'superjson'
+import SuperJSON from "superjson"
 
-import { useState } from 'react'
+import { useState } from "react"
+import { QueryClientProvider } from "@tanstack/react-query"
 
-import { QueryClientProvider } from '@tanstack/react-query'
-import type { QueryClient } from '@tanstack/react-query'
+// types
+import type { QueryClient } from "@tanstack/react-query"
 
 // trpc
-import { httpBatchLink } from '@trpc/client'
+import { httpBatchLink } from "@trpc/client"
 
-import { api as trpc } from '@/trpc/client'
-import { makeQueryClient } from '@/trpc/query-client'
+import { trpc } from "@/server/trpc/client"
+import { makeQueryClient } from "@/server/trpc/query-client"
 
 let clientQueryClientSingleton: QueryClient
 function getQueryClient() {
-  if (typeof window === 'undefined') {
+  if (typeof window === "undefined") {
     return makeQueryClient()
   }
   return (clientQueryClientSingleton ??= makeQueryClient())
@@ -23,9 +24,9 @@ function getQueryClient() {
 
 function getUrl() {
   const base = (() => {
-    if (typeof window !== 'undefined') return ''
-    if (process.env.NODE_ENV !== 'development') return process.env.NEXT_PUBLIC_SERVER_URL
-    return 'http://localhost:3000'
+    if (typeof window !== "undefined") return ""
+    if (process.env.NODE_ENV !== "development") return process.env.NEXT_PUBLIC_SERVER_URL
+    return "http://localhost:3000"
   })()
   return `${base}/api/trpc`
 }
@@ -46,9 +47,7 @@ const TRPCProvider = (props: Readonly<{ children: React.ReactNode }>) => {
 
   return (
     <trpc.Provider client={trpcClient} queryClient={queryClient}>
-      <QueryClientProvider client={queryClient}>
-        {props.children}
-      </QueryClientProvider>
+      <QueryClientProvider client={queryClient} {...props} />
     </trpc.Provider>
   )
 }
