+++ src/server/trpc/client.ts
@@ -1,0 +1,6 @@
+// trpc
+import { createTRPCReact } from "@trpc/react-query"
+
+import { TRPCRouter } from "@/server/trpc/router"
+
+export const trpc = createTRPCReact<TRPCRouter>()
