+++ src/server/trpc/query-client.ts
@@ -1,9 +1,8 @@
 import SuperJSON from "superjson"
-
 import {
   defaultShouldDehydrateQuery,
   QueryClient
-} from '@tanstack/react-query'
+} from "@tanstack/react-query"
 
 export function makeQueryClient() {
   return new QueryClient({
